gplazma: add homedir plugin
This patch adds support for the homedir gPlazma plugin, which provides support for
checking that a user's home directory exists in dCache's namespace.

The homedir plugin can choose a user's home directory.

If a plugin configured earlier in the session chain has already created the home
directory attribute then that path is used.  If not then the plugin is configured
either to fail or to chose home directory by appending the username to a fixed path.

The homedir plugin also verifies the existence of the home directory.

If the path exists but is not owned by the user's uid or is not a directory then the
plugin will fail.  The group-ownership of the directory is not checked.

If the home directory path doesn't exist then the plugin may be configured to fail or
to attempt to create the directory.  It will create the directory with the user's uid
as owner, the user's primary gid as the group-owner, and configurable permissions
("mode bits").  This requires that the parent directory already exists.