Add ability to choose factory method for creating an argument
Currently, the Command method for CLI uses a hard-coded "valueOf" method when
attempting to converting the supplied String to the type of the argument.
While this works in many cases, there are case where a different factory method
is desirable.

This patch adds support for the 'factory' argument to the Argument and Option
annotation.  The default value is 'valueOf', so existing code is not affected.
By specifying the 'factory' argument, the correct factory method may be specified.