vorolemap: fix some bugs and add support for dn-only mappings
There are several problems with the current implementation of the
VoRoleMap plugin.

First, the plugin can add the same groupname as both primary and
non-primary.  This might or might not result in problems with
subsequent plugins, but it is counter-intuitive.

Second, the empty-string FQAN "" is treated as a wildcard.  If this
mapping appears before a dn+fqan mapping line then lines after this
will be ignored.

Third, the plugin fails to parse non-FQAN lines; e.g. "dn groupname".

Fourth, if there are no successful match mapping lines in the vorolemap
for the primary FQAN (or the parent FQANs of the primary FQAN) but
subsequent FQANs provide a group then plugin does not create any
primary group.

This patch fixes these issues.  The new semantics are describe in the
Javadoc for the VoRoleMap class, but in summary:

  1. the plugin maps the primary FQAN (and parents thereof, if
     unsuccessful), then other FQANs are mapping, then finally the DN
     is mapped by itself.

  2. the first successful mapping is primary, all others are
     non-primary.

  3. DN-only mapping lines are processed and only match the final
     DN-only mapping.

  4. The empty FQAN ("") is treated as if its a DN-only mapping line.

Some of the unit-tests had to be updated.  Some because they assert
that "" is a wildcard FQAN that matches any entry (no longer the case),
some because they assert incorrect behaviour (see FIXME comments) and
some because they assume that a DN-only mapping should match an empty
string; whereas now it matches a null.