srm: fix file ownership queries
Currently, the SRM relies on down-casting the supplied SRMUser object to an
AuthorizationRecord to discover the uid and gid of this user.  This is used either
when querying whether a file is owned by some user or the user is a member of the
group that owns this file.

Since we recently moved away from AuthorizationRecord and used DcacheUser as the
concrete class that provides SRMUser, the ability to discover uid and gid of a user
was lost.

This patch fixes the problem by downcasting to the correct class.

A problem not fixed by this patch is that group-ownership is checked only against
the user's primary gid (all groups should be checked).  A future patch will fix
this.

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/5905
Acked-by: Dmitry Litvintsev
Acked-by: Gerd Behrmann