cells: add start method
The cells package currently does not provide a mechanism for a cell to start
activity outside of the constructor.  This results in various hacks to try
to ensure that threads started in the constructor do not see partially
created objects.  It also leads to error messages being erronously
reported as bugs.

This patch fixes these problems by introducing a start method.