poolmanager: provide reasonable error message if partition type is unknown
If the admin has saved a partition type 'classic' in their
poolmanager.conf then dCache will fail to start with a meaningless
error message and a stack-trace.

This patch provides a meaningful error message.  Due to a
hack/work-around for another problem, the message is wrapped in a
RuntimeException, which triggers reporting the problem as a bug (with
a stack-trace).

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/5920/
Acked-by: Karsten Schwank