webdav: support querying only the file attributes needed
The milton library supports PROPFIND queries to discover custom
properties; however, the ResourceFactory is given no mechanism to
discover which custom properties will be requested.  Currently the
factory requests all properties, which is clearly wasteful for the
common case of querying items in a directory.

This patch adds support so that only the additional file attributes
needed for the requested custom properties are requested from
PnfsManager.

Target: master
Requires-notes: no
Requires-book: no