configuration: fix one-of checking
The one-of annotation restricts a value to one of a set of values.
When written, the one-of support didn't consider that the value,
when set, might not be a literal value but contain a reference to
some other property.  However, this is common procedure when
deprecating a property.

The work-around was to extend the set of valid values to include
the reference to the deprecated property.  This patch moves this
work-around into the code: if a property value is a single
expansion then the validity of the data is not checked.

Note that the value cannot be expanded as, during dCache startup,
it isn't guaranteed that the target property has been loaded.

Target: master