systemtest: add support for generating user credentials
Add support for the --user option.  This will use the disposable
CA to generate user credentials in addition to the host credentials.

The existing generated certificate's CN and subjectAltName (email) are
used as default CN and subjectAltName, if present, otherwise the git
'user.name' and 'user.email' config options are used.  Either may be
overwritten by specifying command-line options.

The script will store these credentials in ~/.globus.  If there is
no existing default credential then it will set the freshly created
credential as the default; but, otherwise, it will leave the existing
default credential alone.

If --generate is specified and existing user credentials are detected
then the --user option is assumed.

Target: master