srm: Minor refactoring of ls command
The 'ls' and 'ls completed' commands are merged, and the 'ls completed'
command is replaced by 'ls -completed=<n>'.

The command line interface is moved from thed dcache-srm module to the
srm-server module and used for both dCache and the standalone server.

List formatting related methods are moved from the SRM class to
SrmCommandLineInterface.

This patch is in anticipation of further extensions to the list command.

Target: trunk
Require-notes: yes
Require-book: yes