srm-client: Add interactive SRM shell
Incomplete proof of concept patch of an interactive srm shell.

In constrast to Paul's claim, I observed that interactive use is MUCH
faster when not having to start a new client instance for every command.
Even when the actual connection to the server is reestablished every
time, the command feels almost instantaneous.

Independently of this patch, I provide a sketch to Paul for how the
server connection could be reused. This would make the shell even faster.

Target: none