namespace: add storage-state FileAttribute
This patch introduces an extra FileAttribute that describes the current
status of the data associated with a regular file.  The are currently
three states:

    BEING_WRITTEN  -- indicates that data is being uploaded

    IMMUTABLE -- indicates that the initial upload has completed.

    LOST -- indicates that the file's content has been lost.

A file only transitions into LOST state from explicit statement
from the administrator.

The FileLocality is a somewhat related concept.  It is expected
that all files that have StorageState.LOST will have
FileLocality.UNAVAILABLE; however, it is possible for files to
be UNAVAILABLE without being LOST (e.g., when all pools containing
a disk-only file's replica are offline).

Previously, FileLocality.LOST existed as a state, but it wasn't
used.