srm: report Job ID-generating DAO errors as IllegalStateException
Currently, if the db-backed JobIdGenerator has a problem with generating
the next int or long ID then the DataAccessException (which is a
subclass of RuntimeException) propagates until it is reported as a bug.

This is bad because it provides too little information for an admin to
investigate and is reported as a bug.

With this patch, such problems are wrapped as an IllegalStateException
that provides additional context information.

Target: master
Request: 2.11
Request: 2.10
Request: 2.9
Request: 2.8
Request: 2.7
Request: 2.6