pool: add disk IO bandwidth monitoring
This patch adds support for monitoring reading or writing to the local
block-storage.  The motivation is a recent problem with transfers
timing out.  After the event, it proved impossible to tell whether the
transfer was slow because the underlying block storage was slow, whether
something remote (remote site or WAN) was to blame.

This patch addresses this issue by creating a monitoring proxy for a
RepositoryChannel that keeps track of how much time is spent waiting
for the underlying filesystem.  This, along with keeping track how long
the client took to connect, allows a site to identify the cause of a
low average bandwidth.

The attributes available to the billing file logging include this
information and two textual descriptions of the corresponding
bandwidth.

NB. this approach will not work as-is for protocols where the mover
is multithreaded (e.g., NFS).

Target: master
Requires-book: yes
Requires-notes: yes