config: add annotation to better describe property usage
This patch adds the concept of a definition context: Global, Domain
and Service.  In the layout file, configuration immediately after
declaring a service have Service context; those immediately after
declaring a domain have Domain context.  All other configuration has
Global context.

This patch replaces the 'not-for-services' annotation with the
annotations 'for-scripts', 'for-domains' and 'for-services' with the
following semantics:

    for-scripts annotates properties that influences scripts
        that operate outside the context of a domain; for
        example, info-provider.

    for-domains annotates properties that influences domains;
        for example, affecting how the JVM is instantiated or
        or cells communication.

    for-services annotates properties that influences one or
        more services.

    It is legal for a property to any combination of the
    for-script, for-domains and for-services annotations.

    As an optimisation, a property that is for-services, not
    for-scripts and not for-domains may be written without the
    for-services annotation; i.e., the 'for-services' is
    assumed if nothing specifies it otherwise.

Although not enforced, it is anticipated that a property that is
referenced by some other property's default value will inherit all
for-* annotations from that other property.

When loading the dCache configuration, the usage is checked against
these annotation, much as not-for-services was checked against Service
context.  A for-scripts or for-domains property configured in a
Service context or a for-scripts property configured in a Domain
context will trigger a warning.

Target: master
Requires-book: yes
Requires-notes: yes