pool: update dcap mover not to use system properties
Previously the dcap mover used system properties to discover
which port or within which port-range it should bind the
socket for incoming (passive) connections.

With this patch, the pool.mover.dcap.port property is respected
from the service context.  If this is zero then the
pool.mover.dcap.port-range property is consulted.

The change allows for pools within the same domain to listen on
different ports.

Target: master
Requires-book: yes
Requires-notes: yes