dcache: extend login-session based authorisation
Currently, in gPlazma, a login session is granted read-only or
read-write access via the ReadOnly attribute.  This allows the gPlazma
session plugin to choose whether the user is allowed to write into
dCache when she logs in.  Despite this option being evaluated for each
login session, the currently supplied gPlazma plugins make rather
limited use of this ability; they always return the same value
(read-only vs read-write) for the same user.

This patch expands the expressiveness of this concept by replacing the
ReadOnly attribute with a login-session restriction policy.  As with
ReadOnly, this authorisation policy augments the permissions contained
in the namespace by reducing the operations the user is allowed to do:
the policy may forbid something that the namespace would allow, but it
never allows something the namespace would forbid.

The two behaviours formerly expressed by the ReadOnly attribute, the
user is only allowed to read data and the user has normal access, are
retained by gPlazma session plugins specifying corresponding policies:
DenyModification and NoRestrictions respectively.

However, by specifying an authorisation policy, a gPlazma plugin has a
far richer set of possible restrictions; for example, a plugin can
restrict which part of the namespace a user will see with this login,
limit downloading to particular files or limit uploading to particular
set of directories.

Target: master
Requires-notes: yes
Requires-book: no