cell: rewrite waitfor command
Motivation:

The waitfor command is available from the System cell and, although it
may be used by service startup scripts, currently no service makes us if
it.  This command is currently very dangerous to use.

Despite the command blocking until the target exists, the command does
not return a Reply object.  The result is that the command blocks the
message handling thread until the target exists: no further commands are
handled until the command finishes.  Depending on the target, this can
take considerable time.  As a consequence, if the target is misspelt or
otherwise never appears then this blocking behaviour continues until the
command times out.

What is worse is that, by default, the command has no time out.  If the
target is mistyped then the command will block the System cell's message
thread until the domain is restarted.

Modification:

The command is rewritten to use a Reply.  Additionally, the Reply
infrastructure is updated to support a CellShell discovering the return
code of a Reply, so updating the internal 'rc' and 'rmsg' variables.

Result:

The waitfor command is safe to use.

Target: master
Requires-book: no
Requires-notes: yes