ftp: add support for SRM cancelling an active upload
Motivation:

Current semantics in dCache is that removing a namespace entry will
cause an upload to fail.  Since the SRM removes the TURL when a PUT
request is aborted, the FTP upload will fail.  However the error message
is vague and cryptic.  Since (for whatever reason) users seem to fixate
on the error reported by the FTP door, this results in SRM induced
transfer failures being difficult for users to diagnose.

Modification:

Introduce a topic that notify doors when an upload will ultimately be
unsuccessful.  PnfsManager is updated to generate such notification on
SRM upload cancellation.

The FTP door is updated to keep track of ongoing uploads.  If it
receives cancel notification about an ongoing upload, it aborts that
upload.  It is updated so that, under such circumstances, it does not
attempt to remove the file from the namespace as that will always fail.

Result:

SRM-managed FTP uploads that cannot succeed due to the SRM request being
aborted will fail faster and with more precise error messages.

Limitations:

The pool is currently unable to distinguish between a
PoolMoverKillMessage triggered by a CancelUploadNotificationMessage and
other triggers, so will attempt to register the file within the
namespace.  Such attempts will always fail, and so be logged by the
pool.  Future patches may address this issue.

This patch addresses only the FTP door; subsequent patches will add
similar support for other protocols.

Target: master
Requires-notes: yes
Requires-book: no