pool: explain why a mover was killed
Motivation:

A mover may be killed by a door or by the pool itself for a variety of
reasons.  Nevertheless, these are all logged both within the door and
within billing identically.  This can and has lead to confusion as to
why the mover must die.

Modification:

Add an optional explanation to the Cancellable interface.

Result:

The motivation behind killing a mover becomes clearer.

Target: master
Requires-notes: yes
Requires-book: no