doors: include explanation when killing mover
Motivation:

Many components outside of the pool, not just the door, may kill a
mover.  These components may kill a mover for a variety of reasons: from
normal operational activity through to bugs, door termination and
incorrect client behaviour.  Currently such activity is logged simply as
killed by door; however, this is not always true and in any case, lacks
potentially useful information as to why the mover was killed.

Modification:

Update the message that kills a mover to include an explanation as to
why this was done.  All components that can kill a mover are updated to
include this information when killing the mover.

Result:

More accurate information is logged when a mover is killed by some
component outside of the pool.

Target: master
Requires-notes: yes
Requires-book: no