vehicles: use FileAttributes when creating namespace entries
Motivation:

When creating an entry the PnfsCreateEntryMessage supports a rigid set
of attributes: owner-uid, owner-gid and mode.  This is already
problematic since, under some circumstances, one wishes to create an
entry without specifying uid, gid, or mode and have the namespace use
some default value.  A place-holder value (-1) is used as a work-around
to describe such circumstances.

With commit a2d66c15, a client may wish to create an entry with specific
ACLs.  Due to the restricted nature of PnfsCreateEntryMessage, a second
request to PnfsManager is required to assign the ACLs to the newly
created object.

Modification:

Use FileAttributes to carry the desired information about the newly
created object.  The desired uid, gid and mode are then specified in
this FileAttributes object.  Additional attributes may also be
specified; the dcap door is updated to specify the ACLs for newly
created objects using this method.  Not all attributes may be specified;
for example, the door cannot create an object with a specific nlink
value.

Result:

Creating a new file with ACLs in the dcap door should be faster.

Target: master
Requires-notes: yes
Requires-book: no