webdav: add logout option
Motivation:

Provide a mechanism to force web-browsers to "loose" cached BASIC
authentication information.  This simulates a "logout", although this
isn't really correct terminology.

Modification:

The AuthenticationHandler has a logout ID.  If the UI navigates the
browser to request a page (that the user must authenticate to view) with
the logout ID then the ID increments and the browser receives a 401
UNAUTHORIZED reponse, even though they successfully authenticated via
BASIC authentication.  This poisons the BASIC cache.  Browsers may
either show this error or prompt the user for fresh credentials.  The ID
allows subsequent attempts to view the same URL to succeed.

This method has the advantage that it works without requiring
JavaScript.

This method has the disadvantage of forcing all clients to refresh their
cached pages after a user has logged out.  This is so they can receive
the new logout ID.  Adding a JavaScript mechanism in addition to this
would mostly alieviate this problem, using this mechanism as a fallback.

Note that the logout page is currently the (ugly) Jetty default page.
This is the result of a separate bug.

The patch fixes the template attributes supplied when generating an
error page with StringTemplate, but does not fix the current
duplication.

The patch also factors out the various uses of servlet request
attributes into a separate helper class.  In some places this results in
a stronger enforcement of stored object type.

Result:

The HTML rendering of webdav has a logout button.

Target: master
Requires-notes: yes
Requires-book: no