ftp: fix broken commit cfa765bb
Motivation:

Previous commit introduced a problem where the ftp door fails to repond
when a broken client attempts to send a command in plain-text.  This
results in the client hanging as it waits for a response that will never
come.

Modification:

Ensure the check is performed once the current request field-member is established.

Result:

The correct response is generated.

Target: master
Requires-notes: no
Requires-book: no