ftp: refactor error handling
Motivation:

The main class handling FTP support is rather large and has evolved over
time with the code sometimes using different mechanisms to achieve the
same goal.

This patch reduces the code size somewhat and consolidates the code to
always throw an exception if there is an error, instead of directly
replying.  Throwing an exception is a better fit for the desired goal: a
single reply that guarantees no further response.

Modification:

FTPCommandException is modified to use Throwable's message as the
human-readable (non-numerical) part of the FTP respones.

The isCommandAllowed method signature is modified to throw an exception
rather than returning a boolean, in keeping with the use of
FTPCommandException for error handling.  The method's name is updated
accordingly.

All error cases are updated to throw an FTPCommandException.  This is
less error-prone than reply-then-return.  In many cases, this change
also results in methods taking up less vertical space, improving on
method readability and cohesion.

Add a checkFTPCommand method, modelled after similar Guava methods, that
is often a much more concised than the equivalent if-then-handle-error.
This further improves method cohesion.

Combined, this results in some 6% reduction in line-count.

Result:

Code that is easier to read and easier to maintain.

Target: master
Requires-notes: no
Requires-book: no