webadmin: dynamically request X509 certificate
Motivation:

The webadmin features two authentication mechanisms: username +
password and X.509 certificate.  These options are available as
different login buttons located on the login page.

Despite allowing the user to choose the authentication mechanism on the
login page, dCache always requests an X.509 certificate when the client
connects.  In at least firefox, this results in the user being prompted
on whether an X.509 credential is to be used and, if so, which
identity.

Modification:

Following a suggestion from Jens Jenson, this patch uses
server-triggered renegotiation to request the client supply an X.509
certificate only after the user has requested an X.509-based login.

Result:

The user is only prompted for any X.509 certificate when the user
chooses to log in via X.509 credential.

Target: master
Require-notes: yes
Require-book: no