gplazma: update multimap to support primary-agnostic matching
Motivation:

When mapping principals, it is sometimes useful to select a rule
independent of whether the group principal (group-name, fqan or gid)
represents a primary group.  Logically, this currently requires
duplicate entries; e.g.,

    fqan:/desy,true         gid:1000
    fqan:/desy,false        gid:1000

    fqan:/desy/dcache,true  gid:1001
    fqan:/desy/dcache,false gid:1002

It is desirable to be able to write a single principal that matches
either primary or non-primary groups; e.g.,

    fqan:/desy         gid:1000
    fqan:/desy/dcache  gid:1001

Although the above simpler configuration currently works, it is achieved
by a hack where all identified fqan principals are considered
non-primary when comparing rules.  This hack means that the following
rule is currently broken and will never fire:

    fqan:/desy,true   gid:2001

Modification:

Introduce the concept of a PrincipalMatcher.  This interface is used to
describe the first element of a rule.  It is a generalisation from
Object#equals and is used when testing if a given principal is mapped.

All non-group principals and any group principals that are declared with
explicit primary (e.g., "fqan:/desy,true") continue to match using
Object#equals.  Group principals that are declared without an explicit
primary field (e.g., "fqan:/desy") require the same name, but will match
either primary or non-primary versions.

The code is heavily refactored to use the new NIO Paths API.

The work-around is removed.

The unit tests are updated to be more consistent with BDD conventions
and generally tidied up considerably.

Result:

Group principals that match either primary or non-primary principals may
be written.

Hopefully, the code should be easier to maintain.

Target: master
Require-notes: no
Require-book: no