vehicles: ensure replyRequired is set if a reply is expected
Motivation:

The Message class includes the replyRequired flag that describes whether
the message is notification like (the sender is not expecting a reply)
or API-call like (where the client expects to learn whether the
operation was successful).

One effect of a message not having the replyRequired flag set is that
errors related to message deliver are NOT reported back to the sender.
For example, space-manager, which proxies pool-manager and pools, does
not return a message to the sender if there is a problem and
replyRequired is not set.

Commit 0b94ca041c updated the doors to use PoolManagerHandler when
communicating with a pool-manager or the pool.  Unlike CellStub, sending
messages with PoolManagerHandler does not automatically set the
replyRequired if the sending cell is expecting a reply.  Several
messages rely on the CellStub behaviour to set replyRequired, so are
sent without replyRequired with this patch.

A concrete effect of this problem is visible when a
PoolAcceptFileMessage targets a space or linkgroup with insufficient
capacity.  Such a request will generate an error but, because the
PoolAcceptFileMessage does not have replyRequire flag set, the error is
not propagated back to the door.  This results in the upload failing
eventually, due to a message timeout and yields a generic timeout error.
This makes such problematic uploads slow (which may introduce further
problems) and difficult to diagnose.

Modification:

This patch considers the replyRequired flag a property of the message,
and not something the message sending framework should set.  Therefore,
the code in CellStub and PnfsHandler that adjust replyRequired are
work-arounds.

Rather than re-implement the work-around in PoolManagerHandler, this
patch fixes the vehicles that are sent without replyRequired, and adds
code to catch these problems.

CellStub is updated to identify messages that are sent expecting a reply
but without setting replyRequired and those messages that are sent
without expecting a reply but with replyRequired set.

Similarly, AbstractCell is updated so that it warns if a class replies
to a Message that didn't have replyRequired set.

Using system-test and the test script, a number of classes have been
identified where a reply is expected but the replyRequired is not set.
These have been fixed.

Result:

Fix a problem where WebDAV uploads that have a space-manager related
problem (such as targeting a reservation with insufficient capacity)
would timeout, rather than returning an error straight away.

In general dCache should react faster to error conditions.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Request: 3.2
Ticket: http://rt.dcache.org/Ticket/Display.html?id=9433
Ticket: http://rt.dcache.org/Ticket/Display.html?id=9244 ?
Require-notes: yes
Require-book: yes